Support function names including '('
Go (golang.org) function names include '(' and ')' characters for functions
representing methods on types. For example, the function named
`net/http.(*liveSwitchReader).Read` is for the "Read" method attached to a
pointer to a "liveSwitchReader" value in the package "net/http".

The output of `perf script` for samples of that function look as follows:

    4f4e61 net/http.(*liveSwitchReader).Read (/path/to/binary)

Go 1.5 and above support full stack call graphs in perf_events when programs
are compiled with a toolchain that's built with GOEXPERIMENT=framepointer.