Support stack collapse for multi-event perf.data files
Consider the perf.data file resulting from the following command:

    perf record -e cycles,instructions,cache-misses

Creating a flamegraph for such a file is currently impossible.
Or, well, it's possible but the resulting file is meaningless and
misleading, as it will aggregate all samples for the different
event types.

This patch adds `--event-filter <event-type>` to create meaningful
flamegraphs from multi-event perf.data files. We can use it like
this:

    perf script | stackcollapse-perf --event-filter cycles

By default, only the samples for the first encountered event type
will be used.